+++ Firmware/IO.cpp
@@ -73,10 +73,10 @@ void ReadButtons()
   }
 }
 
-int adc_mv[MAX_ADC];
-int chan_mv[MAX_CHANNELS];
-int chan_ms[MAX_CHANNELS];
-int chan_pct[MAX_CHANNELS];
+int16_t adc_mv[MAX_ADC];
+int16_t chan_mv[MAX_CHANNELS];
+int16_t chan_ms[MAX_CHANNELS];
+int16_t chan_pct[MAX_CHANNELS];
 int nb_adc = 0;
 
 void ReadValues() {
@@ -191,7 +191,7 @@ void ProcessValues() {
     
     
     float us_to_pct = 100.0f/((float)(Config::ConfigFile.channels[i].max_us - Config::ConfigFile.channels[i].min_us));
-    chan_pct[i] = (int)((chan_ms[i]-Config::ConfigFile.channels[i].min_us)*us_to_pct);
+    chan_pct[i] = (int16_t)((chan_ms[i]-Config::ConfigFile.channels[i].min_us)*us_to_pct);
 
     // Set PPM value
     ppmEncoder.setChannel(i, chan_ms[i]);
